ml/pyproject.toml
[tool.poetry]
name = "trixieverse-ml"
version = "0.1.0"
description = "Predictive rank climbing ML pipeline for TrixieVerse"
authors = ["TrixieVerse Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.3.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
aiohttp = "^3.9.0"
asyncio = "^3.4.3"
psycopg2-binary = "^2.9.0"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
requests = "^2.31.0"
joblib = "^1.3.0"
torch = "^2.1.0"
torch-geometric = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
